src/a2a_mcp/requirements.txt
google-adk==0.1.0
mcp[cli]
python-dotenv
starlette
uvicorn
pydantic
diff-match-patch